Start of a level editor
Laid some foundations for a level editor.
F1 to toggle edit mode
Resources can now be invalidated to force a reload (RemoveValue)
Children can be removed from nodes in Engine.h
Keys are now validated in the resource managers. Used it to load an error texture if the file doesn't exist.